# src/main/resources/swagger.yaml
swagger: "2.0"
info:
  version: 1.0.0
  title: User Exercises Rest API
  license:
    name: Aaron Williams
host: localhost:8080
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
paths:
  /api/exercises:
    get:
      summary: List all exercises
      operationId: returnAllExercises
      tags:
        - Exercises
      responses:
        "200":
          description: A paged array of exercises
          schema:
            $ref: '#/definitions/Exercises'
    post:
      summary: Create an exercise
      operationId: createExercises
      tags:
       - Exercises
      parameters:
        - name: exercise
          in: body
          description: Exercise to be created
          required: true
          schema:
            $ref: "#/definitions/Exercise"
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/Exercise'
  /api/exercises/{id}:
    get:
      summary: Info for a specific exercise
      operationId: returnExerciseById
      tags:
        - Exercises
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the exercise to retrieve
          type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Exercise'
...
definitions:
  Exercise:
    required:
      - id
      - description
    properties:
      id:
        type: integer
        format: int64
      description:
        type: string
  Exercises:
        type: array
        items:
          allOf:
            - $ref: '#/definitions/Exercise'
...